refactor(p2p): split hathor.p2p.peer_discovery into separate modules
